move helper packages to pkg/
